User detail 적용

+++ src/main/java/server/nanum/domain/UserGroup.java
@@ -17,5 +17,5 @@ public class UserGroup {
     private Long id;
 
     @Column(name = "point")
-    private Long point;
+    private int point;
 }
